Makefile, equal, iterator

+++ include/vector/sequence_access.h
@@ -19,7 +19,7 @@ namespace ft
 
 	template <typename T, class Al>
 	typename vector<T, Al>::reference vector<T, Al>::at(size_type n) {
-		if (n >= Last - First)
+		if (n >= (u_int64_t)(Last - First))
 			throw std::out_of_range("vector");
 		return (*(begin() + n));
 	}
